% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fkmeans.R
\name{fkmeans}
\alias{fkmeans}
\title{K-Means Clustering for functional data}
\usage{
fkmeans(
  fdataobj,
  ncl = 2,
  metric = metric.lp,
  dfunc = func.trim.FM,
  max.iter = 100,
  par.metric = NULL,
  par.dfunc = list(trim = 0.05),
  method = "sample",
  cluster.size = 5,
  draw = TRUE,
  ...
)
}
\arguments{
\item{fdataobj}{\code{\link{fdata}} class object.}

\item{ncl}{See details section.}

\item{metric}{Metric function, by default \code{\link{metric.lp}}.}

\item{dfunc}{Type of depth measure, by default FM depth.}

\item{max.iter}{Maximum number of iterations for the detection of centers.}

\item{par.metric}{List of arguments to pass to the \code{metric} function.}

\item{par.dfunc}{List of arguments to pass to the \code{dfunc} function .}

\item{method}{Method for selecting initial centers. If
\code{method}=\emph{"Sample"} (by default) takes \code{n} times a random
selection by the \code{ncl} centers. The \code{ncl} curves with greater
distance are the initial centers. If \code{method}=\emph{"Exact"} calculated
all combinations (if < 1e+6) of \code{ncl} centers. The \code{ncl} curves with greater
distance are the initial centers (this method may be too slow).}

\item{cluster.size}{Minimum cluster size (by default is 5). If a cluster has fewer curves,
it is eliminated and the process is continued with a less cluster.}

\item{draw}{=TRUE, draw the curves in the color of the centers.}

\item{\dots}{Further arguments passed to or from other methods.}
}
\value{
Return:
\itemize{
 \item \code{cluster}: Indexes of groups assigned.
\item \code{centers}: Curves centers.
}
}
\description{
Perform k-means clustering on functional data.
}
\details{
The method searches the locations around which are grouped data (for a
predetermined number of groups).\cr

If \code{ncl=NULL}, randomizes the initial centers, \code{ncl=2} using
\code{kmeans.center.ini} function.\cr If \code{ncl} is an integer,
indicating the number of groups to classify,\cr are selected \code{ncl}
initial centers using \code{kmeans.center.ini} function.\cr If \code{ncl} is
a vector of integers, indicating the position of the initial centers with
\code{length(ncl)} equal to number of groups.\cr If \code{ncl} is a
\code{fdata} class objecct, \code{ncl} are the initial centers curves with
\code{nrow(ncl)} number of groups.\cr
}
\examples{
\dontrun{
library(fda.usc)
data(phoneme)
ind <- c(1:50,101:150,201:250)
clas <- as.numeric(phoneme$classlearn[ind])
fdataobj <- phoneme$learn[ind,]
# Unsupervised classification
out.fd1 <- fkmeans(fdataobj,ncl=3,draw=TRUE)
out.fd2 <- fkmeans(fdataobj,ncl=3,draw=TRUE,method="exact")
# Different Depth function
ind2 <- c(17,77,126)
out.fd3 <- fkmeans(fdataobj,ncl=fdataobj[ind2,],draw=FALSE,
                     dfunc =func.trim.FM,par.dfunc=list(trim=0.1))
out.fd4 <- fkmeans(fdataobj, ncl=fdataobj[ind2,], 
                     draw=FALSE, dfunc = func.med.FM)
table(clas,out.fd1$cluster )
table(clas,out.fd4$cluster )
}

}
\references{
Hartigan, J. A. and Wong, M. A. (1979). \emph{A K-means
clustering algorithm}. Applied Statistics 28, 100 \-108.
}
\seealso{
See Also generic \link[stats]{kmeans} function.
}
\author{
Manuel Febrero-Bande, Manuel Oviedo de la Fuente
\email{manuel.oviedo@udc.es}
}
\keyword{cluster}
