% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfkmeans.R
\name{mfkmeans}
\alias{mfkmeans}
\title{K-Means Clustering for Multivariate Functional Data}
\usage{
mfkmeans(mfdata, ncl = 2, max.iter = 100, tol = 1e-04)
}
\arguments{
\item{mfdata}{A list of `fdata` objects, or an `mfdata` object (a list of `fdata` objects) representing multiple functional datasets.}

\item{ncl}{Number of clusters.}

\item{max.iter}{Maximum number of iterations.}

\item{tol}{Tolerance for convergence.}
}
\value{
A list with the following components:
\describe{
  \item{\code{cluster}}{A vector of cluster assignments for each observation.}
  \item{\code{centers}}{A list of centroids for each cluster for each functional variable.}
  \item{\code{iter}}{Number of iterations until convergence.}
}
}
\description{
This function performs k-means clustering on multivariate functional datasets using distances calculated with `metric.mfdata`.
}
\examples{
data(aemet, package = "fda.usc")
datos <- mfdata("temp"=aemet$temp,"logprec"=aemet$logprec)
dd = metric.mfdata(datos)
result <- mfkmeans(datos, ncl = 3)
plot(datos,col=result$cluster)
# plot(aemet$df[,7:8],col=result$cluster,asp=T)

}
\keyword{cluster}
