% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rproc2clust.R
\name{rproc2clust}
\alias{rproc2clust}
\title{Simulate Functional Data for Clustering}
\usage{
rproc2clust(
  n.1 = 51,
  n.2 = n.1,
  mu.1,
  mu.2,
  s.1 = 1,
  s.2 = 1,
  par.list.1 = FALSE,
  par.list.2 = FALSE,
  ...
)
}
\arguments{
\item{n.1, n.2}{Integers specifying the number of curves to be generated 
in the first  and second groups, respectively.}

\item{mu.1, mu.2}{Objects of class \code{fdata} representing the theoretical 
mean functions for the first and second groups, respectively.}

\item{s.1, s.2}{Numeric values specifying the standard deviation of the noise 
to be added to the first and second groups, respectively. 
See the \code{sigma} argument of the \code{rproc2fdata} function.}

\item{par.list.1, par.list.2}{Optional lists of additional parameters to be 
passed to \code{rproc2fdata} for generating the first and second groups, respectively. 
If set to \code{FALSE}, default parameter values will be used.}

\item{...}{Additional arguments to be passed to \code{rproc2fdata}, allowing 
for further customization of the process generation.}
}
\value{
A list containing the following elements:
\itemize{
  \item \code{X}: The functional data as an \code{fdata} object.
  \item \code{X.hat}: The estimated mean functions for each group.
  \item \code{group}: The theoretical functions for each group.
  \item \code{groups}: The group labels for each curve.
  \item \code{color}: The colors associated with each group.
  \item \code{colors}: The colors assigned to each individual curve.
}
}
\description{
Simulates functional data based on specified parameters for 
clustering analysis within the `fda.clust` package. It uses `rproc2mu` 
to generate the latent processes and `rproc2fdata` from `fda.usc` to generate
 the functional data.
}
\examples{
fproces <- rproc2mu( 51, "sin", "cos", 1, 1)
res <- rproc2clust (10,50,fproces$mu.1,fproces$mu.2)
plot(res$X, col=res$groups)
lines(res$X.hat, lwd=2, col=res$color)

fproces <- rproc2mu( 51, "sin", "sin", 1, 2)
res <- rproc2clust (10,50,fproces$mu.1,fproces$mu.2)
plot(res$X, col=res$groups)
lines(res$X.hat, lwd=2, col=res$color)

fproces <- rproc2mu( 101, "poly1", "poly1", 10, -10, 1, 1)
res <- rproc2clust (10,50,fproces$mu.1,fproces$mu.2)
plot(res$X, col=res$colors)
lines(res$X.hat, lwd=2, col=res$color)

fproces <- rproc2mu( 101, "poly1", "poly1", 10, -10, 1, 1)
res <- rproc2clust (10,50,fproces$mu.1,fproces$mu.2,"vexponential","vexponential")
plot(res$X, col=res$colors)
lines(res$X.hat, lwd=2, col=res$color)

fproces <- rproc2mu( 101, "poly1", "poly2", -3, 3, 1, 1.4)
res <- rproc2clust (50,50,fproces$mu.1,fproces$mu.2,"brownian","brownian")
plot(res$X, col=res$colors)
lines(res$X.hat, lwd=2, col=res$color)

par.list.1 <- list(scale = 0.8, theta = 0.3, H = 0.5)
par.list.2 <- list(scale = 0.2, theta = 0.3, H = 0.5)
fproces <- rproc2mu( 101, "poly1", "poly2", -3, 3, 1, 1.4)
res <- rproc2clust (50,50,fproces$mu.1,fproces$mu.2,
                    "brownian","brownian", par.list.1, par.list.2)
plot(res$X, col=res$colors)
lines(res$X.hat, lwd=2, col=res$color)
}
\keyword{cluster}
