% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprocKclust.R
\name{rprocKclust}
\alias{rprocKclust}
\title{Simulate Functional Data for K Clusters}
\usage{
rprocKclust(t, n, process, c, k = NULL, s = 1, par.list = NULL, ...)
}
\arguments{
\item{t}{A numeric vector representing the evaluation points (\code{argvals}) 
at which the functional data are computed.}

\item{n}{A vector of integers specifying the number of curves to be generated 
for each of the K clusters.}

\item{process}{A character vector of length K specifying the type of each latent process.
Options include \code{"poly1"}, \code{"poly2"}, \code{"sin"}, \code{"cos"}, 
\code{"sin_cos"}, and \code{"cos_sin"}.}

\item{c}{A numeric vector of length K specifying the coefficients for each process.}

\item{k}{An optional numeric vector of length K specifying the exponents for the 
processes. This parameter is only relevant for \code{"poly1"} and \code{"poly2"}.}

\item{s}{A numeric vector specifying the standard deviation of the noise to be 
added to each cluster's data. If only one value is provided, it is recycled for all clusters.}

\item{par.list}{A list of lists of additional parameters to be 
passed to \code{rproc2fdata} for generating the functional data for each cluster.}

\item{...}{Additional arguments to be passed to \code{rproc2fdata}, allowing 
for further customization of the process generation.}
}
\value{
A list containing the following elements:
\itemize{
  \item \code{X}: The functional data as an \code{fdata} object.
  \item \code{X.hat}: The estimated mean functions for each cluster.
  \item \code{groups}: The group labels for each curve.
  \item \code{color}: The colors associated with each cluster.
  \item \code{colors}: The colors assigned to each individual curve.
}
}
\description{
Simulates functional data for K clusters based on specified parameters for 
clustering analysis within the `fda.clust` package. It uses `rprocKmu` 
to generate the latent processes and `rproc2fdata` from `fda.usc` to generate
the functional data.
}
\examples{
t <- seq(0, 2*pi, length.out = 101)
res <- rprocKclust(t, 
                   n = c(30, 50, 40), 
                   process = c("poly1", "sin", "cos"), 
                   c = c(10, 1, 1), 
                   k = c(2, NA, NA), 
                   s = c(0.2, 0.3, 0.1))

plot(res$X, col = res$colors)
lines(res$X.hat, lwd = 2, col = res$color)

}
\keyword{cluster}
