% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprocKmu.R
\name{rprocKmu}
\alias{rprocKmu}
\title{Generate Latent Processes for Functional Data}
\usage{
rprocKmu(t, process, c, k = NULL)
}
\arguments{
\item{t}{A numeric vector specifying the grid points (\code{argvals}) at which 
the latent processes are evaluated.}

\item{process}{A character vector specifying the types of the latent processes. 
Available options include \code{"poly1"}, \code{"poly2"}, \code{"sin"}, 
\code{"cos"}, \code{"sin_cos"}, and \code{"cos_sin"}.}

\item{c}{A numeric vector of coefficients used to scale each of the latent processes.}

\item{k}{An optional numeric vector of exponents used in the latent processes. 
These exponents control the shape and curvature of the processes, 
and are only applicable for \code{"poly1"} and \code{"poly2"} process types.}
}
\value{
A list containing the following elements:
\itemize{
  \item \code{mu}: A list of \code{fdata} objects, each representing one of the 
  latent processes.
}
}
\description{
Generates multiple latent processes (\code{mu.1}, \code{mu.2}, ..., \code{mu.k}) 
based on the specified input parameters, including process type, coefficients, exponents, 
and the number of evaluation points.
}
\details{
The available latent process types are described as follows: 
\itemize{
  \item \code{"poly1"}: \eqn{\mu_i(t) = c_i \cdot (1 - t) \cdot t^{k_i}}, 
  where \code{c.i} is the scaling coefficient and \code{k.i} controls the curvature.
  
  \item \code{"poly2"}: \eqn{\mu_i(t) = c_i \cdot (1 - t)^{k_i} \cdot t}, 
  where \code{c.i} is the scaling coefficient and \code{k.i} controls the curvature.
  
  \item \code{"sin"}: \eqn{\mu_i(t) = c_i \cdot \sin(t)}, 
  where \code{c.i} is the scaling coefficient that adjusts the amplitude.
  
  \item \code{"cos"}: \eqn{\mu_i(t) = c_i \cdot \cos(t)}, 
  where \code{c.i} is the scaling coefficient that adjusts the amplitude.
  
  \item \code{"sin_cos"}: \eqn{\mu_i(t) = c_i \cdot (\sin(t) - \cos(t))}, 
  where \code{c.i} is the scaling coefficient.
  
  \item \code{"cos_sin"}: \eqn{\mu_i(t) = c_i \cdot (\cos(t) - \sin(t))}, 
  where \code{c.i} is the scaling coefficient.
}
}
\examples{
t <- seq(0, 1, length.out = 101)
fproces <- rprocKmu(t, c("poly1", "poly2", "sin", "cos"), 
                    c(10, -10, 1, 1), c(1, 1, NA, NA))
plot(fproces, main = "Latent Processes")                   

}
\keyword{cluster}
